--- configs/requirements.txt
@@ -1,1 +1,0 @@
-requests==2.20.0
