+++ src/aws-glue/joblib/common/constants.py
@@ -1,0 +1,1 @@
+A = "A"
