+++ src/aws-glue/joblib/variables.py
@@ -1,0 +1,8 @@
+NAMESPACE='ai4e-ap-southeast-1-dev'
+ARTIFACT_BUCKET_NAME='ai4e-ap-southeast-1-dev-s3-code-artifact'
+DATA_LANDING_BUCKET_NAME='ai4e-ap-southeast-1-dev-s3-data-landing'
+RAW_ZONE='raw_zone'
+BRONZE_ZONE = 'bronze_zone'
+SILVER_ZONE = 'silver_zone'
+GOLDEN_ZONE='golden_zone'
+INSIGHT_ZOME = 'insight_zone'
