+++ src/aws-lambda/requirements.txt
@@ -1,0 +1,1 @@
+requests==2.20.0
